It's amazing
I think most of the code is done -- there still needs to be a few
adjustments, but the code works pretty smoothly. Arduino processes the
accelerometer data, and if the device changed position, it's going to
send that data to processing. Processing will take that data and move
the mouse position accordingly.